+++ cspell.json
@@ -14,6 +14,8 @@
     "conovel",
     "Corrected",
     "Declaration",
+    "docs",
+    "emotion",
     "Ensure",
     "Gutters",
     "icons",
@@ -21,6 +23,7 @@
     "Importing",
     "isolated",
     "learn",
+    "linear",
     "Linting",
     "Locals",
     "logo",
@@ -28,10 +31,18 @@
     "logos",
     "more",
     "pages",
+    "prefers",
+    "prepare",
     "prettier",
+    "prop",
+    "react",
+    "reduced",
     "Remy",
     "responsive",
     "skip",
+    "spin",
+    "staged",
+    "styled",
     "stylelint",
     "Switch",
     "Syntheric",
