+++ src/App.css
@@ -5,16 +5,16 @@
   text-align: center;
 }
 
-.logo {
+.App-logo {
   height: 6em;
   padding: 1.5em;
   will-change: filter;
   transition: filter 300ms;
 }
-.logo:hover {
+.App-logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
-.logo.react:hover {
+.App-logo.react:hover {
   filter: drop-shadow(0 0 2em #61dafbaa);
 }
