+++ tsconfig.json
@@ -20,5 +20,5 @@
     "types": ["node"],
     "resolveJsonModule": true
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite.config.ts", "./src/**/*", "src/custom.d.ts"]
 }
